Split out periodic background functions
