+++ server/bleep/src/periodic.rs
@@ -1,0 +1,5 @@
+mod logrotate;
+mod remotes;
+
+pub(crate) use logrotate::*;
+pub(crate) use remotes::*;
