+++ server/bleep/src/periodic/logrotate.rs
@@ -1,0 +1,3 @@
+pub(crate) fn log_and_branch_rotate(clone: crate::Application) -> _ {
+    todo!()
+}
